refactore: remove shared components and logging middleware
- Deleted shared API and configuration files to streamline the codebase.
- Removed logging middleware and associated logging functionality.
- Eliminated AI service implementation and related endpoints.
- Introduced new Ollama API endpoints for health checks and translation services.
- Updated core configuration and logger to enhance logging capabilities.
- Added new middleware structure for improved request/response logging.
