+++ apps/server/src/api/router.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter, HTTPException, Query
-from src.shared.services.ai import AIService
+from src.services.ai import AIService
 from fastapi.responses import JSONResponse
 from pathlib import Path
 from typing import Optional
@@ -25,4 +25,3 @@ async def get_ai_health():
 async def get_ai_models():
     service = AIService()
     return await service.list_models()
-
