+++ apps/server/src/middlewares/cors.py
