--- apps/server/src/routers/__init__.py
@@ -1,1 +1,0 @@
-"""Routers package."""
