+++ apps/server/src/services/ai.py
@@ -2,7 +2,7 @@ from typing import Optional, List, Any
 from openai import OpenAI
 # from langchain_openai import ChatOpenAI
 from openai.types import Model
-from src.shared.config import get_ai_settings
+from src.core.config import get_ai_settings
 from datetime import datetime, timedelta
 
 class AIService:
